added warnings, removed xl support, added cast to head off sign-conversion error

+++ lib/mpigrid.c
@@ -50,7 +50,7 @@ int getTaskID(const char f, int * argc, char ** argv){
       /* Hide these elements from future invocations of getopt(3). */
       memmove(&argv[optind_prev],
 	      &argv[optind],
-	      (*argc - optind) * sizeof(char *));
+	      (size_t) (*argc - optind) * sizeof(char *));
       *argc -= optind - optind_prev;
 
       /* Stop after this round. */
